move config,schema,logger out of internal
