+++ api/handlers/task_handler.go
@@ -7,9 +7,9 @@ import (
 	"github.com/go-playground/validator/v10"
 	"github.com/google/uuid"
 	"github.com/labstack/echo/v4"
-	"github.com/surendratiwari3/paota/internal/logger"
-	"github.com/surendratiwari3/paota/internal/schema"
-	"github.com/surendratiwari3/paota/internal/schema/errors"
+	"github.com/surendratiwari3/paota/logger"
+	"github.com/surendratiwari3/paota/schema"
+	"github.com/surendratiwari3/paota/schema/errors"
 	"github.com/surendratiwari3/paota/workerpool"
 	"io/ioutil"
 	"net/http"
