+++ api/main.go
@@ -4,8 +4,8 @@ import (
 	"github.com/labstack/echo/v4"
 	"github.com/sirupsen/logrus"
 	"github.com/surendratiwari3/paota/api/handlers"
-	"github.com/surendratiwari3/paota/internal/config"
-	"github.com/surendratiwari3/paota/internal/logger"
+	"github.com/surendratiwari3/paota/config"
+	"github.com/surendratiwari3/paota/logger"
 )
 
 func main() {
