+++ bench/amqpnostorage/consumer/benchmark_test.go
@@ -2,8 +2,8 @@ package consumer
 
 import (
 	"github.com/surendratiwari3/paota/bench/amqpnostorage"
-	"github.com/surendratiwari3/paota/internal/logger"
-	"github.com/surendratiwari3/paota/internal/schema"
+	"github.com/surendratiwari3/paota/logger"
+	"github.com/surendratiwari3/paota/schema"
 	"github.com/surendratiwari3/paota/workerpool"
 	"testing"
 	"time"
