+++ bench/amqpnostorage/setup.go
@@ -3,8 +3,8 @@ package amqpnostorage
 import (
 	"context"
 	"github.com/sirupsen/logrus"
-	"github.com/surendratiwari3/paota/internal/config"
-	"github.com/surendratiwari3/paota/internal/logger"
+	"github.com/surendratiwari3/paota/config"
+	"github.com/surendratiwari3/paota/logger"
 	"github.com/surendratiwari3/paota/workerpool"
 	"os"
 )
