+++ config/amqp.go
