+++ example/publisher/main.go
@@ -4,13 +4,13 @@ import (
 	"context"
 	"encoding/json"
 	"github.com/sirupsen/logrus"
-	"github.com/surendratiwari3/paota/internal/config"
-	"github.com/surendratiwari3/paota/internal/schema"
+	"github.com/surendratiwari3/paota/config"
+	"github.com/surendratiwari3/paota/schema"
 	"github.com/surendratiwari3/paota/workerpool"
 	"sync"
 
 	//"github.com/surendratiwari3/paota/example/task"
-	"github.com/surendratiwari3/paota/internal/logger"
+	"github.com/surendratiwari3/paota/logger"
 	"os"
 )
