+++ internal/broker/amqp/broker.go
@@ -5,13 +5,13 @@ import (
 	"encoding/json"
 	"fmt"
 	amqp "github.com/rabbitmq/amqp091-go"
+	"github.com/surendratiwari3/paota/config"
 	"github.com/surendratiwari3/paota/internal/broker"
-	"github.com/surendratiwari3/paota/internal/config"
-	"github.com/surendratiwari3/paota/internal/logger"
 	"github.com/surendratiwari3/paota/internal/provider"
-	"github.com/surendratiwari3/paota/internal/schema"
-	"github.com/surendratiwari3/paota/internal/schema/errors"
 	"github.com/surendratiwari3/paota/internal/workergroup"
+	"github.com/surendratiwari3/paota/logger"
+	"github.com/surendratiwari3/paota/schema"
+	"github.com/surendratiwari3/paota/schema/errors"
 	"sync"
 	"time"
 )
