+++ schema/amqp.go
