+++ schema/errors/broker.go
@@ -4,5 +4,5 @@ import "errors"
 
 var (
 	ErrUnsupportedBroker = errors.New("unsupported broker")
-	ErrEmptyMessage      = errors.New("Received an empty message")
+	ErrEmptyMessage      = errors.New("received an empty message")
 )
