+++ schema/errors/connection.go
