+++ schema/errors/store.go
