+++ schema/errors/worker.go
