+++ workerpool/workerpool.go
@@ -5,13 +5,13 @@ import (
 	"errors"
 	"fmt"
 	"github.com/google/uuid"
+	"github.com/surendratiwari3/paota/config"
 	"github.com/surendratiwari3/paota/internal/broker"
-	"github.com/surendratiwari3/paota/internal/config"
 	"github.com/surendratiwari3/paota/internal/factory"
-	"github.com/surendratiwari3/paota/internal/logger"
-	"github.com/surendratiwari3/paota/internal/schema"
 	"github.com/surendratiwari3/paota/internal/store"
 	"github.com/surendratiwari3/paota/internal/task"
+	"github.com/surendratiwari3/paota/logger"
+	"github.com/surendratiwari3/paota/schema"
 
 	"github.com/surendratiwari3/paota/internal/workergroup"
 	"os"
