+++ workerpool/workerpool_test.go
@@ -5,11 +5,11 @@ import (
 	"errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
+	"github.com/surendratiwari3/paota/config"
 	"github.com/surendratiwari3/paota/internal/broker"
-	"github.com/surendratiwari3/paota/internal/config"
 	"github.com/surendratiwari3/paota/internal/factory"
-	"github.com/surendratiwari3/paota/internal/schema"
 	"github.com/surendratiwari3/paota/internal/task"
+	"github.com/surendratiwari3/paota/schema"
 	"os"
 	"testing"
 	"time"
